requirements.txt
Django>=6.0.1
djangorestframework
django-unfold[all]
django-cors-headers
